Split the recommender class into the manager and logic
